requirements/base.txt
# Core
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3

# Database
psycopg2-binary==2.9.10

# Authentication
django-allauth==65.3.0
dj-rest-auth==7.0.0
djangorestframework-simplejwt==5.4.0

# Payments
stripe==11.4.1

# Environment
django-environ==0.11.2

# CORS
django-cors-headers==4.6.0

# Task Queue
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Storage
django-storages==1.14.4
boto3==1.35.81

# Utilities
Pillow==11.0.0
django-extensions==3.2.3
django-model-utils==5.0.0

# API Documentation
drf-spectacular==0.28.0

# Security
django-csp==3.8
django-ratelimit==4.1.0
